Revert "Move "check" decorators to the plasmapy.utils.decorators package (#647)"
This reverts commit 9c9c5f680d0c90676fe5c862362c493d1ac36443.
