+++ docs/utils/index.rst
@@ -28,5 +28,6 @@ Reference/API
 =============
 
 .. automodapi:: plasmapy.utils.decorators
+.. automodapi:: plasmapy.utils.checks
 .. automodapi:: plasmapy.utils.exceptions
 .. automodapi:: plasmapy.utils.pytest_helpers
