+++ plasmapy/atomic/ionization_states.py
@@ -2,18 +2,19 @@
 A class for storing ionization state data for multiple elements or
 isotopes.
 """
-__all__ = ["IonizationStates"]
 
+from numbers import Real, Integral
+from typing import Dict, List, Optional, Tuple, Union
 import collections
-import numpy as np
 
-from astropy import units as u
-from numbers import (Real, Integral)
-from plasmapy.atomic import (atomic_number, Particle, particle_symbol, IonizationState,
-                             State)
-from plasmapy.atomic.exceptions import (AtomicError, ChargeError, InvalidParticleError)
+import numpy as np
+import astropy.units as u
+
+from plasmapy.atomic import atomic_number, Particle, particle_symbol, IonizationState, State
 from plasmapy.utils import check_quantity
-from typing import (Dict, List, Optional, Tuple, Union)
+from plasmapy.atomic.exceptions import AtomicError, ChargeError, InvalidParticleError
+
+__all__ = ["IonizationStates"]
 
 
 class IonizationStates:
