+++ plasmapy/utils/decorators/__init__.py
@@ -1,7 +1,6 @@
 """
 A module to contain various decorators used to build readable and useful code.
 """
-__all__ = ['check_relativistic', 'check_quantity', 'preserve_signature']
+__all__ = ['preserve_signature']
 
 from .helpers import preserve_signature
-from .checks import (check_relativistic, check_quantity)
